Move FontContext to fontique
